+++ app/Http/Controllers/ProjectController.php
@@ -4,9 +4,12 @@ namespace App\Http\Controllers;
 
 use App\Project;
 use App\Client;
+use App\Work;
 use App\User;
 use Illuminate\Http\Request;
 use Log;
+use Redirect;
+use App\UserProject;
 
 class ProjectController extends Controller
 {
@@ -51,7 +54,7 @@ class ProjectController extends Controller
             'name'                  => 'required',
             'description'           => 'required|string|max:144',
             'start_date'            => 'required',
-            'expected_end_date'     => 'required|after_or_equal:start_date',
+            'expected_end_date'     => 'required|after:start_date',
             'cost_pr_hour'          => 'required',            
         ]);
 
@@ -60,7 +63,7 @@ class ProjectController extends Controller
         $project->description = $input['description'];
         $project->start_date = $input['start_date'];
         $project->expected_end_date = $input['expected_end_date'];
-        $project->effective_end_date = $input['effective_end_date'];
+        //$project->effective_end_date = $input['effective_end_date'];
         $project->cost_pr_hour = $input['cost_pr_hour'];
         $project->note = $input['note'];
         $project->client_id = $input['client_id'];
@@ -81,6 +84,8 @@ class ProjectController extends Controller
     {
         LOG::info($project);
 
+        
+        //$project->expected_end_date = $request->filled('effective_end_date') ? date('Y-m-d H:i:s', strtotime($request->input('effective_end_date'))) : NULL;
         return view('projects.show', compact('project'));
 
     }
@@ -93,7 +98,7 @@ class ProjectController extends Controller
      */
     public function edit(Project $project)
     {
-        Log::info($project);
+        //Log::info($project);
         return view('projects.edit', compact('project'));
     }
 
@@ -109,14 +114,28 @@ class ProjectController extends Controller
         $input = $request->all();
 
         $validatedData = $request->validate([  
-            'client_id'             => 'required',      
+                 
             'name'                  => 'required',
             'description'           => 'required|string|max:144',
             'start_date'            => 'required',
-            'expected_end_date'     => 'required|after_or_equal:start_date',
-            'effective_end_date'    => 'required|after_or_equal:start_date',
+            'expected_end_date'     => 'required|after:start_date',
             'cost_pr_hour'          => 'required|numeric',
         ]);
+
+        /* Controllo data fine effettiva, se viene inserita*/
+        if($request->effective_end_date > 0){
+            $request->validate([ 'effective_end_date'    => 'sometimes|after:start_date', ]);
+            
+            $query = Work::join('projects','projects.id','works.project_id')->where('projects.name',$request->name)
+            
+                        ->where('date','>',$request->effective_end_date)
+                        ->get();
+            //LOG::info($query);
+            
+            if($query->isNotEmpty()){
+                return Redirect::back()->withErrors("Data Fine Effettiva, ci sono attività svolte dopo il $request->effective_end_date");
+            }
+        }
 
         $project->name = $input['name'];
         $project->description = $input['description'];
@@ -138,8 +157,17 @@ class ProjectController extends Controller
      */
     public function destroy(Project $project)
     {
-        $project->delete();
+        /* Controllo che il progetto non sia associato ad utenti o ad attività */
+        $associazioni = UserProject::where('project_id',$project->id)->get(); 
+        $attivita = Work::where('project_id',$project->id)->get(); 
+        
+        if($associazioni->isNotEmpty() || $attivita->isNotEmpty()){
+            LOG::info($associazioni);
+            return view('projects.edit', compact('project'))->withErrors("il progetto è associato ad utenti o attività");
+        }
 
+        $project->delete();
+        
         return redirect('/admin/project');
     }
 }
